# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_merka_api_internal_domain.Permissao:
    enum:
    - criar_usuario
    - criar_perfil
    - configurar_sistema
    - ver_auditoria
    - ver_relatorios
    - lancar_item
    - remover_item
    - registrar_peso
    - estornar_peso
    - transferir_mesa
    - aplicar_desconto
    - cancelar_comanda
    - processar_pagamento
    - entregar_comanda
    - cadastrar_produto
    - configurar_preco_peso
    - cancelar_nota_fiscal
    type: string
    x-enum-varnames:
    - PermissaoCriarUsuario
    - PermissaoCriarPerfil
    - PermissaoConfigurarSistema
    - PermissaoVerAuditoria
    - PermissaoVerRelatorios
    - PermissaoLancarItem
    - PermissaoRemoverItem
    - PermissaoRegistrarPeso
    - PermissaoEstornarPeso
    - PermissaoTransferirMesa
    - PermissaoAplicarDesconto
    - PermissaoCancelarComanda
    - PermissaoProcessarPagamento
    - PermissaoEntregarComanda
    - PermissaoCadastrarProduto
    - PermissaoConfigurarPrecoPeso
    - PermissaoCancelarNotaFiscal
  github_com_merka_api_internal_domain.PermissionCatalogo:
    properties:
      chave:
        $ref: '#/definitions/github_com_merka_api_internal_domain.Permissao'
      descricao:
        type: string
      id:
        type: string
    type: object
...
  github_com_merka_api_internal_domain.Role:
    properties:
      id:
        type: string
      nome:
        type: string
      sistema:
        type: boolean
      tenantID:
        type: string
    type: object
...
  internal_handler.criarPerfilRequest:
    properties:
      nome:
        type: string
      permissoes:
        items:
          type: string
        type: array
    type: object
...
  internal_handler.editarPermissoesPerfilRequest:
    properties:
      permissoes:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Sistema de comandas multi-tenant (churrascaria como primeira instância).
    Ver CLAUDE.md e docs/merka-planejamento.md para o contexto completo do domínio.
  title: Merka API
  version: "1.0"
paths:
...
  /perfis:
    get:
      description: Admin Super e Gestor podem ver — ver comentário em RegistrarRotas
        sobre a permissão reaproveitada.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_merka_api_internal_domain.Role'
            type: array
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar perfis do tenant
      tags:
      - perfis
    post:
      consumes:
      - application/json
      description: Restrito a Admin Super (permissão "criar_perfil"). Recebe nome
        e uma lista de chaves de permissão já existentes no catálogo — cria o role
        e as linhas correspondentes em role_permissions.
      parameters:
      - description: Nome do perfil e chaves de permissão
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.criarPerfilRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_merka_api_internal_domain.Role'
        "400":
          description: nome obrigatório ou permissão inexistente no catálogo
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: já existe um perfil com esse nome
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Criar perfil customizado (US-02)
      tags:
      - perfis
  /perfis/{id}/permissoes:
    put:
      consumes:
      - application/json
      description: 'Restrito a Admin Super (permissão "criar_perfil"). Bloqueia perfis
        de sistema (ex: "Admin Super") — imutáveis, pra não travar o próprio acesso
        do sistema.'
      parameters:
      - description: ID do perfil
        in: path
        name: id
        required: true
        type: string
      - description: Novo conjunto de permissões
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.editarPermissoesPerfilRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_merka_api_internal_domain.Role'
        "400":
          description: permissão inexistente no catálogo
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: perfil não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: perfil de sistema é imutável
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Substituir as permissões de um perfil (US-02)
      tags:
      - perfis
  /permissoes:
    get:
      description: Admin Super e Gestor podem ver — mesma permissão reaproveitada
        de ListarPerfis.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_merka_api_internal_domain.PermissionCatalogo'
            type: array
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar catálogo fixo de permissões
      tags:
      - perfis
...
securityDefinitions:
  BearerAuth:
    description: Informe "Bearer <token>" (token obtido em POST /auth/login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"